add confirm before deletion

+++ client/src/app/components/list/list.component.ts
@@ -1,10 +1,11 @@
 import { Component, ViewChild, OnInit } from '@angular/core';
 import { first } from 'rxjs/operators';
 import { Router } from '@angular/router';
-import { MatSnackBar, MatPaginator, MatSort, MatTableDataSource } from '@angular/material';
+import { MatSnackBar, MatPaginator, MatSort, MatTableDataSource, MatDialogConfig, MatDialog } from '@angular/material';
 
 import { LoginService, AuthenticationService } from '../../_services';
 import { Login, User } from '../../_models';
+import { DialogComponent } from '../../_components';
 
 @Component({
   selector: 'app-list',
@@ -19,7 +20,7 @@ export class ListComponent implements OnInit {
   @ViewChild(MatSort, {static: false}) sort: MatSort;
 
   constructor(
-    private authenticationService: AuthenticationService,
+    private authenticationService: AuthenticationService, private dialog: MatDialog,
     private loginService: LoginService, private router: Router, private snackBar: MatSnackBar) {
       this.currentUser = this.authenticationService.currentUserValue;
     }
@@ -48,17 +49,31 @@ export class ListComponent implements OnInit {
     this.router.navigate([`/edit/${id}`]);
   }
 
-  deleteLogin(id) {
-    this.loginService.delete(id).subscribe(() => {
-      this.snackBar.open('login deleted successfully', 'OK', {
-        duration: 3000,
-      });
-      this.fetchLogins();
-    }, (error) => {
-      this.snackBar.open(error, 'Try again', {
-        duration: 3000,
-      });
-    })
+  deleteLogin(id, comments) {
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+
+    dialogConfig.data = {
+        id: 1,
+        description: 'Do you really want to delete this?: ' + comments
+    };
+
+    const dialogRef = this.dialog.open(DialogComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe(data => {
+      if (data) {
+        this.loginService.delete(id).subscribe(() => {
+          this.snackBar.open('login deleted successfully', 'OK', {
+            duration: 3000,
+          });
+          this.fetchLogins();
+        }, (error) => {
+          this.snackBar.open(error, 'Try again', {
+            duration: 3000,
+          });
+        })
+      }
+    });
   }
 
   applyFilter(filterValue: string) {
